[bugfix]兼容gradle3.0以上组件独立运行时出现的问题：https://github.com/luckybilly/CC/issues/62

+++ cc-register/src/main/groovy/com/billy/android/register/RegisterPlugin.groovy
@@ -21,6 +21,7 @@ public class RegisterPlugin implements Plugin<Project> {
         println "project(${project.name}) apply ${PLUGIN_NAME} plugin"
         project.extensions.create(EXT_NAME, RegisterExtension)
         def isApp = ProjectModuleManager.manageModule(project)
+        performBuildTypeCache(project, isApp)
         if (isApp) {
             println "project(${project.name}) register ${PLUGIN_NAME} transform"
             def android = project.extensions.getByType(AppExtension)
@@ -32,9 +33,15 @@ public class RegisterPlugin implements Plugin<Project> {
                     ManifestGenerator.generateManifestFileContent(project, config.excludeProcessNames)
                 }
             }
-        } else {
+        }
+    }
+
+    private static void performBuildTypeCache(Project project, boolean isApp) {
+        if (!RegisterCache.isSameAsLastBuildType(project, isApp)) {
+            RegisterCache.cacheBuildType(project, isApp)
             //兼容gradle3.0以上组件独立运行时出现的问题：https://github.com/luckybilly/CC/issues/62
-            def cachedJniFile = project.file("build/intermediates/transforms/mergeJniLibs")
+            //切换app/lib编译时，将transform目录清除
+            def cachedJniFile = project.file("build/intermediates/transforms/")
             if (cachedJniFile && cachedJniFile.exists() && cachedJniFile.isDirectory()) {
                 FileUtils.deleteDirectory(cachedJniFile)
             }
